backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pymongo==4.6.0
python-decouple==3.8
Pillow==10.1.0
django-filter==23.5
PyJWT==2.8.0
twilio==8.10.0
sendgrid==6.11.0
python-dotenv==1.0.0
anthropic
openai
gTTS
openai-whisper
torch
torchvision
torchaudio
